ENSITECH-285 split types into themes and component files
