+++ packages/tabs/src/components/ShowMore/index.tsx
@@ -10,7 +10,7 @@ import {
 import type { FC, FocusEvent, MouseEvent } from 'react';
 
 import { useTabsTheme } from '../../context';
-import type { ShowMoreButtonProps } from '../../types';
+import type { ShowMoreButtonProps } from '../../types/component';
 
 // eslint-disable-next-line @typescript-eslint/no-unused-vars
 const ButtonField = ({
