+++ packages/tabs/src/components/Title/LinkTitle.tsx
@@ -2,7 +2,7 @@ import Link, { type LinkProps } from 'next/link';
 import { type HTMLProps, forwardRef } from 'react';
 
 import { useTabsTheme } from '../../context';
-import type { TabListTitle } from '../../types';
+import type { TabListTitle } from '../../types/component';
 
 type Props = TabListTitle & Omit<LinkProps, 'passHref'> & HTMLProps<HTMLAnchorElement>;
