+++ packages/tabs/src/themes/basic/index.ts
@@ -1,8 +1,6 @@
 import { type OptionizedCSS, defaultTheme, extractCSSOption, scale } from '@ensi-platform/core-components-common';
 
-// eslint-disable-next-line import/no-cycle
-import type { TabsSize, TabsTheme } from '../../types';
-// eslint-disable-next-line import/no-cycle
+import type { TabsSize, TabsTheme } from '../../types/themes';
 import { toggle } from './toggle';
 
 const { colors, typography } = defaultTheme;
