+++ packages/tabs/src/themes/basic/toggle.ts
@@ -1,7 +1,7 @@
 import { type OptionizedCSS, defaultTheme, extractCSSOption, scale } from '@ensi-platform/core-components-common';
 
 // eslint-disable-next-line import/no-cycle
-import type { TabsSize, TabsTheme } from '../../types';
+import type { TabsSize, TabsTheme } from '../../types/themes';
 
 const { colors, typography } = defaultTheme;
