+++ packages/tabs/src/themes/index.ts
@@ -1,5 +1,4 @@
-// eslint-disable-next-line import/no-cycle
-import type { TabsTheme } from '../types';
+import type { TabsTheme } from '../types/themes';
 import { basicTheme } from './basic';
 
 export const TABS_THEMES = {
